refactor: 서버-클라, 증권사-서버 패키지 분리
