+++ websocket/src/main/java/com/fintory/websocket/provider/config/WebSocketClientConfig.java
@@ -1,7 +1,7 @@
-package com.fintory.websocket.config;
+package com.fintory.websocket.provider.config;
 
-import com.fintory.websocket.handler.KoreanLiveStockPriceWebSocketHandler;
-import com.fintory.websocket.handler.OverseasLiveStockPriceWebSocketHandler;
+import com.fintory.websocket.provider.handler.KoreanLiveStockPriceWebSocketHandler;
+import com.fintory.websocket.provider.handler.OverseasLiveStockPriceWebSocketHandler;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
