+++ websocket/src/main/java/com/fintory/websocket/provider/config/WebSocketInterceptor.java
@@ -1,4 +1,4 @@
-package com.fintory.websocket.config;
+package com.fintory.websocket.provider.config;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.server.ServerHttpRequest;
